DNN-8299 Changes for lockfile deletion rights.
Parallel programming with ash.

+++ DNN Platform/Library/Services/Installer/Blocker/InstallBlocker.cs
@@ -24,31 +24,35 @@ namespace DotNetNuke.Services.Installer.Blocker
         #endregion
 
         #region Members
+        private bool fileCreated = false;
         #endregion
 
         #region Public Methods
 
         public void RegisterInstallBegining()
-        {
-            File.Create(Globals.ApplicationMapPath + installBlockerFile);
+        {          
+            if(!fileCreated)  
+                File.Create(Globals.ApplicationMapPath + installBlockerFile);
+            fileCreated = true;
         }
 
         public bool IsInstallInProgress()
         {
-            return File.Exists(Globals.ApplicationMapPath + installBlockerFile);
+            return fileCreated || File.Exists(Globals.ApplicationMapPath + installBlockerFile);
         }
 
         public void RegisterInstallEnd()
         {
             var retryable = new RetryableAction(() =>
             {
-                if (IsInstallInProgress())
+                if (IsInstallInProgress() && fileCreated)
                 {
                     File.Delete(Globals.ApplicationMapPath + installBlockerFile);
                 }
             }, "Deleting lock file", 60, TimeSpan.FromSeconds(1));
 
             retryable.TryIt();
+            fileCreated = false;
         }
 
         #endregion
